feat(portal): 添加门户访问统计功能
- 修改数据库表结构将PV和UV字段从INT改为BIGINT类型
- 新增PortalVisit实体类用于访问统计数据映射
- 创建PortalVisitController提供访问上报和统计查询接口
- 实现PortalVisitService业务逻辑处理访问统计
- 添加VisitConverter转换器处理数据传输对象映射
- 定义VisitTotalDTO和VisitTotalResponse数据传输对象
- 实现访问上报的原子操作和累计统计查询功能

+++ src/main/resources/sql/init_ddl.sql
@@ -530,8 +530,8 @@ CREATE TABLE IF NOT EXISTS `portal_footprint` (
 CREATE TABLE IF NOT EXISTS `portal_visit` (
     `id`          BIGINT     NOT NULL                COMMENT '主键ID',
     `stat_date`   DATE       NOT NULL                COMMENT '统计日期',
-    `pv`          INT        NOT NULL DEFAULT 0      COMMENT '当日访问量（PV）',
-    `uv`          INT        NOT NULL DEFAULT 0      COMMENT '当日独立访客数（UV）',
+    `pv`          BIGINT     NOT NULL DEFAULT 0      COMMENT '当日访问量（PV）',
+    `uv`          BIGINT     NOT NULL DEFAULT 0      COMMENT '当日独立访客数（UV）',
     `creator`     BIGINT         NULL DEFAULT NULL       COMMENT '创建人',
     `create_time` DATETIME   NULL DEFAULT CURRENT_TIMESTAMP       COMMENT '创建时间',
     `updater`     BIGINT         NULL DEFAULT NULL       COMMENT '更新人',
